requirements.txt
attrs==24.2.0
certifi==2024.8.30
exceptiongroup==1.2.2
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
PySocks==1.7.1
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
pytest~=8.3.3
allure-pytest==2.13.5
allure-python-commons==2.13.5

pip~=24.2
cffi~=1.15.1
pluggy~=1.5.0
iniconfig~=2.0.0
websockets~=13.1
requests~=2.32.3
Faker~=30.8.2